texcache-rewrite: Virtual EFB copies for OpenGL.
We still need a depalettizing shader for OpenGL.

git-svn-id: https://dolphin-emu.googlecode.com/svn/branches/texcache-rewrite@7549 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Plugins/Plugin_VideoDX11/Src/PSTextureEncoder.cpp
@@ -487,7 +487,7 @@ PSTextureEncoder::~PSTextureEncoder()
 	SAFE_RELEASE(m_outRTV);
 }
 
-size_t PSTextureEncoder::Encode(u8* dst, unsigned int dstFormat, D3DTexture2D* srcTex,
+u32 PSTextureEncoder::Encode(u8* dst, unsigned int dstFormat, D3DTexture2D* srcTex,
 	unsigned int srcFormat, const EFBRectangle& srcRect, bool isIntensity,
 	bool scaleByHalf)
 {
