+++ Source/Plugins/Plugin_VideoDX11/Src/PSTextureEncoder.h
@@ -30,7 +30,7 @@ public:
 	PSTextureEncoder();
 	~PSTextureEncoder();
 
-	size_t Encode(u8* dst, unsigned int dstFormat, D3DTexture2D* srcTex,
+	u32 Encode(u8* dst, unsigned int dstFormat, D3DTexture2D* srcTex,
 		unsigned int srcFormat, const EFBRectangle& srcRect, bool isIntensity,
 		bool scaleByHalf);
